Good Chunk of Code with Some Clear Errors
My GitHub hasn't really been working, so this is a lot of stuff at
once. I know there are some errors and I have commented in some of the
places I need to fix. I'm trying to figure out how to get two different
speeds:a current speed and an initial speed that is 0.25 seconds behind
the current speed.